Make it possible to define components without decorator
This makes it possible to simply assign a Component to a variable
and just call it again to get a new Component instance, merging
keyword arguments with the previous instance kwargs.
This requires all components to only accept keyword-arguments,
because positional arguments (tuple) could not be merged, and
the rules would be complicated. This is explicit, easy to remember
and implement, and is probably more flexible.