Only set children in the same calling frame
It's more intuitive that parent tracking with context managers does not
continue over separate components. We can determine the current context
managers in the same component by looking at the parent/previous/calling frame
(frame.f_back). If it's the same, there is a "frame boundary" which is more
intuitive to follow. The problem is the global ContextVar, which is set
no matter where a new context manager starts. This way we can "localize"
the parent to the same component/function/object.

See test for example.