core: Don't keep None defaults
None is probably the only "value" referring
to something as missing. This makes it possible
to dynamically calculate component arguments
and just pass None when we want to leave out
and argument.

Falsy values are totally different though,
we want to keep empty string, because we can't
know if that's intended or meant to be a "not-present"
value. If that's the case, None should be used as default
and changed to empty string if necessary.