core: use inspect.Signature instead of custom implementation
It's in the standard library, does everything needed for properly
passing parameters to functions and raises TypeErrors when needed.
Probably better to use it than re-implementing my own.
The only gotcha is that I don't think that BoundArguments.arguments
works properly when mutating it, we need to work around that.

Also fixed some bad tests.