core: Fix merging props
using bound_args.arguments.update() was not correct
when signature contained **kwargs. Be more precise
by only updating the proper parts of bound_args.

Also fixed the props property by checking for **kwargs in signature.

convert args from props in _HTMLComponentBase, this way merging works
and _convert_kwargs is simpler. Also only evaluate attributes when
actually needed (on rendering). The only downside for this is that
Exceptions will be raised only then instead of when __init__ is called.