core: calling partial calls .replace()
instead of .merge(). It is more common to
want to replace something. It is more verbose to call
a partial and immediately .replace() something else.

This also better mimics a Component with defaults.
When instantiating a Component, defaults will be overwritten.
This way, it's less difference between a partial and a real
Component.