core: Simplify API
Deleted Partial Component, as it is not required at all.
Python builtin functools.partial can be used, it does
almost everything what I implemented in it.

Allow __call__ of Components. It makes the same as instantiating
a component instance or calling a partial. Doesn't need to make
a difference, because the result is usually clear when calling;
"update this component attributes". One can still use .replace()
if they want to be more precise, but I think, most of the time
it's not even neccessary. Might make sense to use it in libraries
or new component frameworks.

Baiscally the public API is now:
(All of these make a new component.)

- Component(*args, **kwargs)
- comp(*arg, **kwargs) -> copy props
- comp.replace(**kwargs) -> replace only existing prop
- comp.append(**kwargs) -> add (+) values to existing prop values