core: Use markupsafe instead of my own escaping
Unfortunately, this requires a dependency, but hopefully
the benefits (e.g. speedup, usage, features) are worth it.

I don't want to vendor it, because I would need to compile C code
and release my own wheels which doesn't worth the cost. Also it's
much easier to upgrade this way and Compone is not a single-file
module anyway.

Fixes #5