core: fix HTML attributes
It is really hard to detect objects which behave like str,
but not subclasses of str.

For example Django's gettext_lazy and reverse_lazy objects are proxies
to objects which have string-like behavior, but there is no way to tell
if they should be treated as strings in a framework-independent way.

The biggest problem is that they are still iterable...
We cannot check for __str__ existence or other protocols
like that, because all objects have __str__.