core: Remove __call__ API
It's error prone, because there is no indication that
a prop was overwritten. Sometimes that's what intended
and that was the point of this API, but it can be achieved
with a better interface I believe. For example, functools.partial
doesn't allow specifying the same parameter twice, which is safer.