core: Check duplicate kwargs
Because of Python keywords are passed with underscore at the end,
and represented as such, it can cause duplicate keyword argument
passed in to Component classes, because e.g. "class" and "class_"
is different. Check these in __init__.

Any other duplicate names will be handled by Signature.bind(),
as that works the same way as calling a Python callable.

Fixes #14