core: raise SyntaxError on Python keywords as arguments
Instead of checking duplicates, don't allow them, as
it would be a SyntaxError in Python if called normally.

This helps keeping things consistent, because we can make sure
every keyword is in underscored form, so no other checks required
anywhere else in the code to test the name.

Might be too strict, will see in real-world usage.