core: benchmarks with pytest-benchmark
For now, it's a microbenchmark against Jinja2 only, doesn't show much,
but it's a start. It already shows that compone is very slow with the
current inspect.Signature based approach.

Both dataclasses and attrs generate source code from string, eval and
apply to the class. That might be faster than Jinja2.

Benchmarks doesn't run by default, have to enable with pytest
--benchmark-enable.